+++ server/src/covet/api/__init__.py
@@ -12,6 +12,7 @@ from covet.api import invitations as invitations_router
 from covet.api import item_templates as item_templates_router
 from covet.api import items as items_router
 from covet.api import loans as loans_router
+from covet.api import maintenance as maintenance_router
 from covet.api import meta as meta_router
 from covet.api import metadata as metadata_router
 from covet.api import share as share_router
@@ -34,5 +35,6 @@ api_router.include_router(invitations_router.router)
 api_router.include_router(audit_router.router)
 api_router.include_router(metadata_router.router)
 api_router.include_router(documents_router.router)
+api_router.include_router(maintenance_router.router)
 
 __all__ = ["api_router"]